+++ chevah/compat/constants.py
@@ -13,7 +13,6 @@ from __future__ import absolute_import
 # http://msdn.microsoft.com/library/windows/desktop/bb762494%28v=vs.85%29.aspx
 CSIDL_FLAG_CREATE = 0x8000
 
-DEFAULT_FILE_MODE = 0o666
 DEFAULT_FOLDER_MODE = 0o777
 
 WINDOWS_PRIMARY_GROUP = u'Users'